requirements.txt
telethon==1.28.5  
python-dotenv==0.21.0